+++ dev1/Views/NewFamilyView.swift
@@ -92,5 +92,5 @@ struct NewFamilyView: View {
 
 #Preview {
     NewFamilyView()
-        .modelContainer(for: [FamilyMember.self, FeeCampaign.self, Payment.self, Expense.self], inMemory: true)
+        .modelContainer(for: [TeamProfile.self, FamilyMember.self, FeeCampaign.self, Payment.self, Expense.self], inMemory: true)
 }